+++ src/main/java/gift/core/domain/product/ProductRepository.java
@@ -1,18 +1,18 @@
-package gift.core.product;
+package gift.core.domain.product;
 
 import jakarta.annotation.Nonnull;
 
 import java.util.List;
 
-public interface ProductService {
+public interface ProductRepository {
 
     Product get(Long id);
 
     boolean exists(Long id);
 
-    void createProduct(@Nonnull Product product);
+    void save(@Nonnull Product product);
 
-    void updateProduct(@Nonnull Product product);
+    int size();
 
     List<Product> findAll();
