+++ src/main/java/gift/core/domain/product/exception/ProductAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package gift.core.product.exception;
+package gift.core.domain.product.exception;
 
 import gift.core.exception.APIException;
 import gift.core.exception.ErrorCode;
